Make sure precomputed intent is not stale (#6572)
closes:
https://linear.app/sourcegraph/issue/CODY-4613/non-deterministic-intent-detection

We precompute the intent while the user is typing the input with a 300ms
debounce. It is possible that after the last intent detection the user
may change the query and submit within 300ms, causing the detected
intent to be stale and often different than the intent for the final
query.

This PR fixes that behavior by checking if the query has changed since
the last intent detection and ignoring the pre-computed intent in case
the query was changed afterwards.


## Test plan

https://sourcegraph.slack.com/archives/C07EQBRSF35/p1735859984632339
## Changelog

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->