feat(webviews): add Sourcegraph Workspaces CTA (#6604)
CLOSE
https://linear.app/sourcegraph/issue/CODY-4661/editor-side-panel-cta

1. The WelcomeNotice component is added to the WelcomeMessage component,
positioned at the bottom of the message. This provides additional
information or notices to the user in the chat welcome message.
2. The isTeamsUpgradeCtaEnabled (feature flag) prop is added to the
WelcomeMessage component, allowing the display of the WelcomeNotice
component to be conditionally rendered based on the value of this prop.
This provides flexibility in controlling the visibility of the
WelcomeNotice component within the chat welcome message.

## Changes
- Added the WelcomeNotice component to the WelcomeMessage component
- Positioned the WelcomeNotice at the bottom of the WelcomeMessage,
centered horizontally
- Added the isTeamsUpgradeCtaEnabled prop to the WelcomeMessage
component
- Conditionally rendered the WelcomeNotice component based on the value
of isTeamsUpgradeCtaEnabled

## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

Log into an account that has the teams upgrade cta enabled feature flag
and confirm you are seeing the CTA in welcome screen:


![image](https://github.com/user-attachments/assets/a3907456-1efb-457d-9d83-91e325e52c1d)

Missing new logo svg