omnibox: add "Did you mean" notice (#6655)
This adds a "did you mean?" notice to the top of chat when a user enters
a query that matches common "search-like" patterns like `search for
useCallback`. When a search-like pattern is detected, we make an LLM
call to extract keywords from the query for use with code search. This
seems to work pretty well in practice, and it avoids the false positive
and "eager LLM" problems from depending on an LLM for actually detecting
whether this is a search-like query.

Closes SRCH-1524

## Test plan

[Video walkthrough](https://share.cleanshot.com/LkDYqsWG)

<table>
<tr>
<td>

![CleanShot 2025-01-15 at 19 15
59@2x](https://github.com/user-attachments/assets/6119b956-16e1-4b4d-b94c-b1923aa001df)



</td>
<td>

![CleanShot 2025-01-15 at 19 16
20@2x](https://github.com/user-attachments/assets/08c07e79-703d-4064-bd79-8da7f7b39715)


</td>
</tr>
</table>