feat: at mentions for prompt templates editor (#6638)
This PR adds a `CodyPromptTemplate` component that will be used in the
prompt templates editor in the prompt library. The first attempt was to
only use the `PromptEditor`, but since that one needs a certain
environment to run it, it was easier to copy the `CodyWebChat` component
and adjust to the needs of the prompt template editor. The component
does not show up in cody web or the editor, and is therefore safe to
merge without a feature flag.

It also adds a `serialize` and `deserialize` which takes @ mention nodes
and turns them into a base64 encoded string. Future iterations may want
to add more nuanced serialization, but this works for now and fits into
our timeline.

See
https://sourcegraph.slack.com/archives/C07HRH4RJP8/p1736809277981609?thread_ts=1736191997.525979&cid=C07HRH4RJP8
for more context on the delivery date.


https://github.com/user-attachments/assets/420e9b14-2663-4dfe-ac5d-de0114329f29

## Test Plan

CI and manual testing through a local SG instance