omnibox: link file path to the line of the first match (#6705)
It's not immediately obvious that the line numbers are links, so users
often end up opening the search result by clicking the file path.
However, the file path doesn't currently link to a line number, just the
beginning of the file. This modifies the link to point to the first line
in the first chunk shown.