feat: Add more control options to cody web (#6715)
This PR makes changes to cody web which are necessary for the new cody
web page updates.

> [!NOTE]
> It's possible that I have to make additional changes to this code as
I'm finalizing the Svelte and React app updates, but the core changes
are ready.

Contributes to
https://linear.app/sourcegraph/issue/DES-267/implement-omnibox-submenu-for-web

### Changes

#### Host / cody web communication

This PR adds a way to have more control over the behaviorTrigger a new
chat. of cody web from the host. This is enabled by cody web
accepting a controller object that has a similar API as the vscode API.
It allows the host and cody web to
send messages to each other.

In particular the host can now do the following:

- Intercept view changes. This allows us to navigate to the native
`/prompts` page instead of cody's prompts panel.
- Get the current transcript. This allows us to show a title on the cody
web page.
- Trigger history deletion and export.
- Start a new chat.

With this abilities we can now implement cody web specific control
elements in the web apps themselves.

#### UI changes

The main change is that it's now possible to render cody web without the
tab bar. The component now accepts a prop to
either render it in 'page' or 'sidebar' mode (cody sidebar will still
show the tabs).

Some additional tweaks have been made according to the new desings, such
as rending the prompts cards in two columns if there
is enough space.

## Test plan

Manual testing in the Svelte app:

- Clicking the history button opens the history panel.
- Clicking the export button downloads the history.
- Clicking the `<-` or `New Thread` buttons starts a new chat
- Clicking on links that would navigate to the prompts panel will load
the prompts page instead.

I'm still finalizing the changes in the Svelte and React apps.


![2025-01-21_12-51_2](https://github.com/user-attachments/assets/bc329589-178c-4d96-a08f-7f21be83a454)

![2025-01-21_12-51_1](https://github.com/user-attachments/assets/9e7a9742-7be5-4d17-a5da-5c04fde64324)

![2025-01-21_12-51](https://github.com/user-attachments/assets/b50380e6-b20c-4208-b5e0-5ad661d29a43)