fix/context: Link to helpful resource when current repo not indexed in non-dotcom (#6695)
When remote search is enabled (that is, Enterprise, etc.) and the
current repo is not indexed remotely:

- Do *not* add a phony "current repo" context item to default context.
- *Do* show an indication that the current repo is not available as
context, and link to a helpful resource. (Part of CODY-4676.)

#6649 make a change to show a "current repo" context item when remote
search is available, but the current repo is not indexed remotely. That
is not right. We want current repo context to be backed by remote search
(although even with remote search there is a local layer on top with
`git status`, we do not support 100% local context ala symf when remote
search is enabled.)

## Test plan

1. Sign in to a dotcom account. Verify that default context, @-mentions
show "Current Repository" and fetch context from it, etc. This behavior
has not changed in this PR and must still work.
2. Sign in to an Enterprise account and open a workspace that *is*
indexed remotely. There should be a "Current Repository" context item by
default; the @-mentions menu should contain "Current Context
github.com/foo", etc. This behavior also has not changed in this PR and
must still work.
3. While still signed in to Enterprise, open a workspace that is *not*
indexed remotely. There should not be a "Current Repository" context
item by default. There should be a new "Current Repository/Not yet
available" menu item, and activating that menu item should open a
browser.
4. Run `pnpm -C web dev` and verify that Cody Web still works as in step
2 above.

![Screenshot 2025-01-20 at 13 17
22](https://github.com/user-attachments/assets/aed6d36b-3fc5-4584-bbc6-373db146af36)