fix(omnibox): add conditional rendering to buttons (#6731)
fixes QA-280

currently we're showing the filter and add to context button even if
there is no search results. clicking on the filter button would still
open up the filters sidebar. this PR introduces a cosmetic cleanup and
will not render these buttons if a user gets no search results

before:
![CleanShot 2025-01-22 at 15 15
31@2x](https://github.com/user-attachments/assets/09f6429f-a853-42a5-95d4-a15b631655fa)


after:
![CleanShot 2025-01-22 at 15 15
59@2x](https://github.com/user-attachments/assets/8e987eb2-1550-4f18-a498-db4bfe28fce9)


note: the padding at the top is missing on main as well


## Test plan
1. type in search query that will generate no results, ex -
asdfjkasdfasdfas
2. filter and "add to context" buttons should not show up
3. type in search query that will generate a search result, double
checked the filters and add to context buttons still work
<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

---------

Co-authored-by: Naman Kumar <naman@outlook.in>