Backport 6833 to jb v7.12.x (#6836)
backport: https://github.com/sourcegraph/cody/pull/6833


This PR:

Removes the intent toggle feature completely.

Changes the intent dropdown behavior to only select the intent and not
execute it automatically.

Enables the submit button even when the chat is empty.

## Test plan


the intent options in the dropdown only selects the intent mode and
doesn't execute it.

after selecting the intent, changing the input, doesn't reset the
intent.

on submit the selected intent is used and the intent selection message
is displayed accordingly.

loading a chat from history and editing the message, doesn't persist the
intent mode previously manually selected.

selecting a prompt correctly executes it in chat & edit modes.

make sure the submit button isn't disabled when the input is empty.


<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->