fix(auto-edit): fix the false notification for auto-edit non eligibility (#6899)
## Context

This PR fixes an issue where users would receive a non-eligibility
notification for auto-edits even when they were in the experiment flag
and it is enabled for them. This happens because of the flaky network
issue, and although we get error while evaluating the feature flag, we
return `false` by default. This creates noisy alerts for the users.

<img width="1054" alt="image"
src="https://github.com/user-attachments/assets/fe14d5e3-1e00-4715-ab59-48693db2249f"
/>

To address this, I added a check that ensures the user is likely viewing
the settings editor (either the JSON settings file or the graphical
Settings UI) before displaying the non-eligibility notification.

## Test plan
1. Limit the networks bandwidth using charles proxy.
2. Switch to the non pro account or override the feature flag
`cody-autoedit-experiment-enabled-flag` to false.
3. You should only get notification if either `settings.json` or the
settings UI is open