fix(context): Mark at-mention context items as user-added items (#6920)
FIX
https://linear.app/sourcegraph/issue/CODY-4843/agentic-context-removes-explicitly-mentioned-remote-file

- Update `getContextForChatMessage` in
`lib/shared/src/context/openctx/context.ts` to set the `source` property
of `ContextItemOpenCtx` to `ContextItemSource.User` to indicate that the
item was added by the user.
- Update `toStructuredMentions` in
`vscode/src/chat/chat-view/ContextRetriever.ts` to set the `source`
property of `ContextItem` to `ContextItemSource.User` if it is not
already set.
- Add a note in `lib/shared/src/codebase-context/messages.ts` explaining
the purpose of the `source` property and how it is used to prioritize
user-added context items.

This change ensures that user-added context items are properly
identified and can be prioritized over non-user-added items when adding
context during the prompt-building process and displaying context in the
chat view.


## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

All the at-mentioned items should have labeled with source = user