allow subscribing to OpenCtx controller before it's initialized
Previously, if some statically initialized value tried to read the `openCtx.controller` before it was set in `exposeOpenCtxClient`, it would throw. This required some contortions to ensure that it was initialized only after `exposeOpenCtxClient` was called. Now, anything can subscribe to `openctxController` at any time, and it will receive an emission of the OpenCtx controller when it's ready.