do not force agent workspace root URI to be `file:` scheme (#6932)
In https://github.com/sourcegraph/cody/pull/5391 the agent started
forcing the workspaceRootUri to `file:` scheme. It was not documented
why, and this is indicative of a bug elsewhere. In the rules PR
(https://github.com/sourcegraph/cody/pull/6909), Cody Web starts caring
about the workspace root URI and needs to use a `repo:` URI scheme.
(Previously, the workspaceRootUri in Cody Web was an invalid URI
`sourcegraph/cody`.)

## Test plan

CI