fix: Stop button on cody web
There are currently two problems with the stop button functionality on
cody web:

1. After pressing the button the response stream continues for a while.
2. The button doesn't turn back into the submit button, making itj
   impossible to submit a follow up question.

This commit solves (2).

Technical details:

The way the chat response propagates through the system somwhat like
this (simplified)

response chunks -> typewriter effect -> webworker transport -> ui

We are also passing an abort signal all the way down to the logic that
initiates the response stream. When the close button is activated we are
trigger that signal.
But this abortion doesn't 'trickle back up' to typewriter, etc. because
`fetchEventSource` doesn't emit any event in this case.

The fix is to directly listen to the abort event and call the completion
callback.