Agent: fix crash when using `TextDocument.validateRange(...)`
Previously, running the agent tests caused the following error to get
logged:
```
----stderr----
Uncaught exception: Error: Method not implemented.
    at _AgentTextDocument.validateRange (/Users/olafurpg/dev/sourcegraph/cody/agent/src/AgentTextDocument.ts:136:15)
    at FixupTask.getContentChanges (/Users/olafurpg/dev/sourcegraph/cody/vscode/src/non-stop/FixupTask.ts:163:41)
```
The root problem is that we are using `TextDocument.validateRange()` in
the VS Code extension and this method is not implemented yet for non-VSC
clients.

I'm not 100% sure how serious this bug is, but the fix is simple enough
that we should just merge the fix.