fix(ui): Use timeout argument directly as duration instead of timestamp (#6954)
When setting the remove timeout for status bar loaders and errors, we
incorrectly treat `StatusBarErrorArgs.timeout` as a timestamp
representing milliseconds since the epoch. However, every usage example
passes the timeout as a number representing how many milliseconds to
wait:

```
this.config.statusBar.addLoader({
    title: 'Completions are being generated',
    timeout: 30_000,
})
```

This PR fixes `StatusBar`'s. `addLoader` and `addError` methods to treat
`StatusBarErrorArgs.timeout` as such.

## Test plan

Local testing via the debugger confirming that the ttl is properly set
to the passed in timeout when it is smaller than one hour.