Use rules (.sourcegraph/*.rule.md) in chat and inline edit (#6909)
For our code review agent (announced in EAP in
https://sourcegraph.com/blog/introducing-enterprise-ai-agents), we're
introducing rules that are `*.rule.md` files placed in `.sourcegraph/`
dirs in your codebase. The code review agent uses these rules to find
and fix issues at code review.

This PR makes it so that Cody in the editor also reads the rules and
uses them in chat and inline edit for higher quality and consistency. It
is behind the `"cody.rules.enabled": true` VS Code user setting, which
is marked as `experimental`.

Docs and blog post coming soon after we've merged this and tested it in
our own dev workflows.

- Closes https://linear.app/sourcegraph/issue/CODY-4829.
- Depends on https://github.com/sourcegraph/sourcegraph/pull/3305.
- Depends on https://github.com/sourcegraph/cody/pull/6932.
- Depends on https://github.com/sourcegraph/cody/pull/6931.
- See related https://github.com/sourcegraph/cody/pull/6930, which makes
rules work in auto-edit.


https://github.com/user-attachments/assets/4d193edc-bdaf-4081-bbf3-c657f91a2bbe


![image](https://github.com/user-attachments/assets/bfee6f0b-1e4d-48d9-8651-8857455f9983)


![image](https://github.com/user-attachments/assets/fb51d77a-4440-4e33-bf3d-55189bc468e3)

## Known issues

- [ ] The `text_content_filters` and `repo_filters` rule fields are not
yet implemented.
- [ ] Editor (not sidebar) chat does not show the rules for the file in
the other pane because it's not active
- [ ] There are Node.js FS errors printed to the debug console when
looking for `.sourcegraph` dirs. These are perhaps not being properly
suppressed.

## Test plan

1. Set `"cody.rules.enabled": true` in your VS Code settings.
1. Clone the https://github.com/sourcegraph/review-agent-sandbox
repository.
1. Open it in VS Code.
1. Open `handler_review.go`.
1. Open Cody chat. Ensure that it shows `N rules` in the chat. Run a
chat and ensure it uses the rules. (It should say so in the response.)
1. Run an inline edit. Ensure that the `Rules` are listed in the inline
edit initially right when you press <kbd>Opt+K</kbd> and then that the
rules apply.
1. Try modifying a rule or creating a new one in
`.sourcegraph/*.rule.md`. Ensure that it starts showing up in chat.

There are no e2e tests yet; this is just being enabled for experimental
dogfooding internally.