fix/release: Update the JetBrains link to release notes. (#6976)
We will create releases in sourcegraph/cody now, because that's where
the commits are! And link to this page to display them:

https://github.com/sourcegraph/cody/releases?q=JetBrains&expanded=true

## Test plan

```
cd jetbrains
./gradlew verifyPlugin
```

Land this, backport it, then release a v7.67 nightly and check that [the
JetBrains Marketplace
page](https://plugins.jetbrains.com/plugin/9682-cody-ai-code-assistant/edit/versions/nightly)
looks good.