chat ui: remove user avatars, remove speaker name from follow-up message
The user knows who they are, why show it next to every message? It
clutters the UI. So I removed the avatar from the human messages (we
still have the avatar in the top right).

I also removed the name from the follow-up message. It doesn't add
anything.

Ideally we'd remove the "user name" every where but when I did that, the
layout looked kinda bad because there was still the 'split into chat'
button.