chat ui: remove user avatars, remove speaker name from follow-up message (#6981)
The user knows who they are, why show it next to every message? It
clutters the UI. So I removed the avatar from the human messages (we
still have the avatar in the top right).

I also removed the name from the follow-up message. It doesn't add
anything.

Ideally we'd remove the "user name" every where but when I did that, the
layout looked kinda bad because there was still the 'split into chat'
button.

## Before

![before](https://github.com/user-attachments/assets/2ce844e1-275a-42e4-876c-5d18bb011a47)

## After

![after](https://github.com/user-attachments/assets/364a23a7-242b-4ed7-a49f-08de8df6b47f)

## Test plan

- N/A