bug fix: Fix 500ms delay when submitting chat message
This removes the 500ms delay when submitting messages. It was
introduced in #6720.

If I understood the PR correctly, it was introduced to fix the error
message that shows up when trying to switch intent while a response is
streaming in.

But instead of waiting 500ms before aborting, we go back to aborting,
and explicitly abort a possible ongoing request before editing the
message and resubmitting it.

In my testing, it fixes the bug.