bug fix: Fix 500ms delay when submitting chat message (#6990)
This removes the 500ms delay when submitting messages. It was introduced
in #6720.

If I understood the PR correctly, it was introduced to fix the error
message that shows up when trying to switch intent while a response is
streaming in.

But instead of waiting 500ms before aborting, we go back to aborting,
and explicitly abort a possible ongoing request before editing the
message and resubmitting it.

In my testing, it fixes the bug.

I also re-added the tests that were commented out in #6720.

## Test plan

- Submit a message to Claude, saying something like "write me a long
poem" so that it generates a lot of text
- While Claude is generating, switch the intent manually by clicking on
the button
- No error message should pop up

(cherry picked from commit aa9c0bcf1dc3bf33eb7be43ba65e0d99bae4be6d)