fix ~5px pixel jitter when pressing Enter in chat in VS Code (#6991)
Previously, pressing Enter in a chat message in VS Code's Cody chat
would cause the iframe to be shifted up by ~5px. This was an annoying
visual jitter. Now, this no longer happens.

## Test plan

CI

Also: Run in VS Code, type a chat message, press Enter, and ensure there
is no jitter. Also ask a long chat question and ensure that the
scroll-to-end works.

## Changelog

- Fixed an issue where pressing Enter in chat would cause brief visual
jitter in the UI.