chore/release: VSCode stable release Slack announcements to point to the correct release branches (#6975)
## Test plan

No test... we will watch what happens in Slack when we release M68 to
stable. This is the last step in the job so if it fails the release is
still done.

The diff URLs look like this now:
https://github.com/sourcegraph/cody/compare/M66...M68

(cherry picked from commit 3afac5d28992f61698be3c3753ff630368b01eac)