disable windows/macOS e2e tests on main, shard ubuntu tests (#7031)
The e2e tests pass fairly reliably on PRs where we break them up into 5
separate runs. But on main we run all the e2e tests in a single go. This
appears to be causing test failures after 9min due to cancellation (and
in general, it is flakier because it just takes 1 e2e test flake).

The Windows and macOS full e2e tests have been failing recently, so they
are disabled. It would be good to reenable them. But now the Ubuntu e2e
tests should pass on main.



## Test plan

n/a