Fix repo name being `Symbol(@@pendingOperation)` if repo is local
Without this fix, when using Cody in a local repo that's not on the
connected instance, this could show up in the prompt:

> Codebase context from file tailscale.py in repository Symbol(@@pendingOperation)

Problem #1 was that the caller of `getRepoNamesContainingUri` didn't
skip the `@@pendingOperation` sentinel value.

Problem #2 was that if we didn't find a repository name on the instance,
we didn't fall back to the local name. We only did that for dot-com
users.