chore: Add prosemirror-lexical interop test
To make the prosemirror implementation work as a drop-in replacement we
use some utility function to convert lexical state to prosemirror and
back.
The utility function was written by carefully inspecting the lexical
serialized state, but that doesn't give us any confidence in its
correctness.

This commit adds a test for these utility functions.