Unclickable links no longer display in Cody WelcomeMessage (#7102)
Before, unclickable prompts were still displayed in the WelcomeMessage
when a user starts a new chat with Cody. This PR makes it so that only
clickable items appear in the PromptList.

The changes enhance the PromptList component by making it more
context-aware with respect to the IDE and prompt modes. This should
provide a better user experience, particularly for web-based clients, by
showing clickable prompts and excluding non-clickable ones.

## Test plan
Before: 
![Screenshot 2025-02-14 at 10 35
10 AM](https://github.com/user-attachments/assets/61e842c6-1473-454c-b448-fc4b3d4777b3)

After: 
![Screenshot 2025-02-14 at 10 35
51 AM](https://github.com/user-attachments/assets/6a6fbbae-c44d-45ba-baef-51786c49b59d)

---------

Co-authored-by: julialeex <julia.xu@sourcegraph.com>