feat(auto-edit): Support unified diff and refactor diff format (#7000)
https://linear.app/sourcegraph/issue/CODY-4921/auto-edit-support-unified-diff-format-for-image-generations


https://github.com/user-attachments/assets/f9a0afdd-7e09-4cd7-85fe-53b99dc1d7fc


## Description

This PR:
- Adds support for a unified diff format with a different UI. This will
be useful for other editors where we do not have support for decorations
or will not have it for the MVP
- Refactors the image generator feature to use `DecorationInfo` instead
of `AddedLinesDecorationInfo`. This means we can easily port this to the
inline decorator in a follow up PR.
- Generally refactors much of the canvas/highlighting/diff logic. This
was necessary to simplify this code as much as possible, as porting to
`DecorationInfo` and supporting deletions here adds a lot of checks/edge
cases.

## Test plan

- Manual testing with both `unified` and `additions` modes enabled
- Unit tests that generate images
- Further tests in this PR branched from here:
https://github.com/sourcegraph/cody/pull/7138

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->